#11 [ADD] SPN 기능 구현
history 에 함수 2개 추가 했음
프로세스 스케줄링이 끝난 마지막에  전력량을 체크하는 방식이던데
나는 1초마다 코어의 전력량을 증가 시키는 쪽으로 구현했음

+++ src/team/os/model/Core.java
@@ -6,7 +6,7 @@ public abstract class Core {
 	// 작업량 PCore = 2, ECore = 1
 	private int power;
 	// 누적 전력 소모량
-	private int accConsumption = 0;
+	private double accConsumption = 0;
 	// 일하고 있는지 여부
 	private boolean working = false;
 	
@@ -29,10 +29,10 @@ public abstract class Core {
 	public void setPower(int power) {
 		this.power = power;
 	}
-	public int getAccConsumption() {
+	public double getAccConsumption() {
 		return accConsumption;
 	}
-	public void setAccConsumption(int accConsumption) {
+	public void setAccConsumption(double accConsumption) {
 		this.accConsumption = accConsumption;
 	}
 }
